// src/views/nolayout/zalo-management/types.ts
export type ZaloAppSettings = {
    appId: string;
    appSecret: string;
    displayName: string;
    appDomain: string;
    contactPhone: string;
    contactEmail: string;
    appIconUrl: string;
    category: string;
    description: string;
    isActive: boolean;
    requireSecretProof: boolean;
};